# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for AI-Dia application",
        "title": "AI-Dia API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8005",
    "basePath": "/api/v1",
    "paths": {
...
        "/client/{client_id}/whatsapp/whatsmeow/disconnect": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Logs out and removes the personal WhatsApp (whatsmeow) session for the tenant.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WhatsApp Whatsmeow"
                ],
                "summary": "Disconnect WhatsApp Whatsmeow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "client_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    }
                }
            }
        },
        "/client/{client_id}/whatsapp/whatsmeow/qr": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "SSE endpoint that streams QR codes for pairing a personal WhatsApp number. Connect with EventSource in the browser. Events: \"qr\" (scan the code), \"connected\" (success), \"timeout\", \"error\".",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "WhatsApp Whatsmeow"
                ],
                "summary": "Stream WhatsApp QR Code (whatsmeow)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "client_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/client/{client_id}/whatsapp/whatsmeow/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns whether the tenant's personal WhatsApp (whatsmeow) is connected.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WhatsApp Whatsmeow"
                ],
                "summary": "WhatsApp Whatsmeow Connection Status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "client_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    }
                }
            }
        },
...
        "/subs/current": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the active subscription status of the currently logged-in tenant.\nReturns active plan info (single/multiple), token usage, and status message.\nIf no active plan exists, returns free plan usage info.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscription"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/subs.Response"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    }
                }
            }
        },
        "/subs/token-usage": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the current AI token usage for the authenticated tenant.\n\n**Free Plan:**\n- `plan_type` = \"free\"\n- `is_unlimited` = false\n- `token_limit` = 1,000,000\n- `tokens_used` = total tokens consumed so far\n- `tokens_remaining` = tokens left before the bot is blocked\n- `percentage_used` = percentage of quota used (0–100)\n- `message` shows a warning when usage ≥ 80%, and an upgrade prompt when limit is reached\n\n**Paid Plan:**\n- `plan_type` = \"paid\"\n- `is_unlimited` = true\n- All numeric fields return -1 (unlimited)\n\n**Error Responses:**\n- 401 — Token missing, expired, or invalid\n- 403 — Authenticated user is not a Client\n- 404 — Tenant record not found\n- 500 — Failed to retrieve subscription or usage data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subscription"
                ],
                "summary": "Get AI Token Usage",
                "responses": {
                    "200": {
                        "description": "Token usage data",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/subs.TokenUsageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized — invalid or missing token",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden — user is not a Client",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Tenant not found",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "helpers.ApiResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "subs.ActivePlanInfo": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "string"
                },
                "plan_name": {
                    "type": "string"
                },
                "tenant_plan_id": {
                    "type": "string"
                },
                "token_usage": {
                    "$ref": "#/definitions/subs.TokenUsage"
                },
                "valid_from": {
                    "type": "string"
                },
                "valid_until": {
                    "type": "string"
                }
            }
        },
        "subs.Response": {
            "type": "object",
            "properties": {
                "active_plans": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/subs.ActivePlanInfo"
                    }
                },
                "current_token_usage": {
                    "$ref": "#/definitions/subs.TokenUsage"
                },
                "is_free": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "subs.TokenUsage": {
            "type": "object",
            "properties": {
                "total_cost": {
                    "type": "number"
                },
                "total_tokens": {
                    "description": "-1 = unlimited",
                    "type": "integer"
                },
                "used_tokens": {
                    "type": "integer"
                }
            }
        },
        "subs.TokenUsageResponse": {
            "type": "object",
            "properties": {
                "is_unlimited": {
                    "description": "true if paid plan active",
                    "type": "boolean"
                },
                "message": {
                    "description": "human-readable summary",
                    "type": "string"
                },
                "percentage_used": {
                    "description": "0-100, -1 if unlimited",
                    "type": "number"
                },
                "plan_type": {
                    "description": "\"free\" | \"paid\"",
                    "type": "string"
                },
                "token_limit": {
                    "description": "1_000_000 for free, -1 for unlimited",
                    "type": "integer"
                },
                "tokens_remaining": {
                    "description": "-1 if unlimited",
                    "type": "integer"
                },
                "tokens_used": {
                    "description": "total tokens consumed",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Format: \"Bearer {token}\" — paste the token from the login response (accessToken)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}